fix: Catch OS Level connection error (#14)
* Remove expicit ping

Removing the ping-handler function and it's call, since websockets sends pings automatically and the function was causing async_io errors.

* Refine error handling and logging

Narrrow websocket exception hadnling to "ConnectionClosedError" from "ConnectionClosed" where feasible.
some logging changes

* expose attribute_map

Expose _attribute_map as public property, since it is used in hass-homee

* Tackle deprecations and lint warnings

- Replace typing.List with list
- remove unnecessary else:s
- use Callable from collections.abc

* refine logging

- log disconnect and reconnect once as per https://developers.home-assistant.io/docs/integration_quality_scale_index/
- use %s instead of f-strings in Log-statements as per HA style guide: https://developers.home-assistant.io/docs/development_guidelines/#use-new-style-string-formatting

* fix typo

* fix some nits

- make sure disconnect warning is only logged once.
- move logging of reconnect attempts to proper location.

* Deal with attributes without options

If trying to access options if the attribute has none leads to an error.
This fixes it.

* Now really deal with empty options

First attempt had a wrong idea.

* Add catch for OS ConnectionError

If a connection error throws at  OS Level (e.g. 113 -  No route to host), it was not caught.